refactor: UI전용 타입 src/types/로 이동
